+++ packages/html-elements/src/elements/__tests__/Breaks.test.ios.tsx
@@ -1,0 +1,10 @@
+import 'react-native';
+import { render } from '@testing-library/react-native';
+import * as React from 'react';
+
+import { BR } from '../Breaks';
+
+it('renders BR', () => {
+  const { toJSON } = render(<BR />);
+  expect(toJSON()).toMatchSnapshot();
+});